feat: add BoardNotFoundException to BoardServiceImpl.

+++ src/main/java/com/example/task_management_system_ampada/services/BoardServiceImpl.java
@@ -1,5 +1,6 @@
 package com.example.task_management_system_ampada.services;
 
+import com.example.task_management_system_ampada.exceptions.BoardNotFoundException;
 import com.example.task_management_system_ampada.models.Board;
 import com.example.task_management_system_ampada.repositories.BoardRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -21,12 +22,19 @@ public class BoardServiceImpl implements BoardService {
 
     @Override
     public Optional<Board> findBoardById(String id) {
-        return boardRepository.findById(id);
+        if (boardRepository.findById(id).isPresent())
+            return boardRepository.findById(id);
+        else
+            throw new BoardNotFoundException();
     }
 
     @Override
     public List<Board> findAllBoards() {
-        return boardRepository.findAll();
+        List<Board> boards = boardRepository.findAll();
+        if (!boards.isEmpty())
+            return boards;
+        else
+            throw new BoardNotFoundException();
     }
 
     @Override
@@ -43,12 +51,15 @@ public class BoardServiceImpl implements BoardService {
             board.setModifiedOn(LocalDateTime.now());
             return boardRepository.save(board);
         }).orElseThrow(() -> {
-            throw new RuntimeException();
+            throw new BoardNotFoundException();
         });
     }
 
     @Override
     public void deleteBoardById(String id) {
-        boardRepository.deleteById(id);
+        if (boardRepository.existsById(id))
+            boardRepository.deleteById(id);
+        else
+            throw new BoardNotFoundException();
     }
 }
